+++ src/Data/Binary/Class.hs
@@ -90,9 +90,7 @@ import Data.Kind (Type)
 import GHC.Exts (RuntimeRep(..), VecCount, VecElem)
 #endif
 import qualified Data.ByteString as B
-#if MIN_VERSION_bytestring(0,10,4)
 import qualified Data.ByteString.Short as BS
-#endif
 import qualified Data.Map        as Map
 import qualified Data.Set        as Set
 import qualified Data.IntMap     as IntMap
@@ -643,12 +641,10 @@ instance Binary ByteString where
     get    = get >>= getLazyByteString
 
 
-#if MIN_VERSION_bytestring(0,10,4)
 instance Binary BS.ShortByteString where
    put bs = put (BS.length bs)
             <> putShortByteString bs
-   get = get >>= fmap BS.toShort . getByteString
-#endif
+   get = get >>= getShortByteString
 
 ------------------------------------------------------------------------
 -- Maps and Sets